Changed it so that the classificaton for images can be changed manually.
It wasn't working before because the CorrectionMenu looks for the
change in the list of results and then puts the selected one at the
front of the list (this is in the select_symbol method). If the new
classification isn't in the list, it would ignore it. Since Imageblobs
only ever have one result, this would never work. I added code to
force it, but this could become obsolete in the future.

git-svn-id: file:///mnt/raid/svn/mathfrontend@109 ba091d2c-d542-4e25-8a58-bd8a2c7d4f9e