Changed the check for iPad to use Modernizr.
Modernizr is a third party library licensed under a BSD/MIT license.
It is used on page load to check for certain features for a browser. This
commit introduces its use for checking for touch events and replaces the earlier
check for user agent. NOTE: This is a costum version of Modernizr built made
on its website. To check for support of features other than touch events, you
must make a new copy and replace the one in this commit.